EVAKA-4455 Move the utilization chart from unit info to unit attendances
