+++ frontend/src/e2e-test/specs/5_employee/units.spec.ts
@@ -70,8 +70,9 @@ describe('Employee - Units', () => {
     })
 
     const unitPage = await unitsPage.nthUnitRow(0).openUnit()
-    await unitPage.assertUnitName(unitFixture.name)
-    await unitPage.assertVisitingAddress(
+    const unitInfoPage = await unitPage.openUnitInformation()
+    await unitInfoPage.assertUnitName(unitFixture.name)
+    await unitInfoPage.assertVisitingAddress(
       `${unitFixture.streetAddress}, ${unitFixture.postalCode} ${unitFixture.postOffice}`
     )
   })
@@ -159,7 +160,8 @@ describe('Employee - Units', () => {
 
   test('Unit occupancy rates cannot be determined when no caretaker', async () => {
     const unitPage = await UnitPage.openUnit(page, unitFixture.id)
-    await unitPage.occupancies.assertNoValidValues()
+    const unitAttendancePage = await unitPage.openAttendancesPage()
+    await unitAttendancePage.occupancies.assertNoValidValues()
   })
 
   test('Unit occupancy rates are correct with properly set caretaker counts', async () => {
@@ -172,8 +174,15 @@ describe('Employee - Units', () => {
       .save()
 
     const unitPage = await UnitPage.openUnit(page, unitFixture.id)
-    await unitPage.occupancies.assertConfirmed('Min. 14,3 %', 'Max. 14,3 %')
-    await unitPage.occupancies.assertPlanned('Min. 14,3 %', 'Max. 14,3 %')
+    const unitAttendancePage = await unitPage.openAttendancesPage()
+    await unitAttendancePage.occupancies.assertConfirmed(
+      'Min. 14,3 %',
+      'Max. 14,3 %'
+    )
+    await unitAttendancePage.occupancies.assertPlanned(
+      'Min. 14,3 %',
+      'Max. 14,3 %'
+    )
   })
 
   test('Units list hides closed units unless toggled to show', async () => {
